moves Data to dimcat.data.base
