+++ src/dimcat/data/base.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+import logging
+
+from dimcat import DimcatObject
+
+logger = logging.getLogger(__name__)
+
+
+class Data(DimcatObject):
+    """
+    This base class unites all classes containing data in some way or another.
+    """
+
+    class Schema(DimcatObject.Schema):
+        # basedir = fields.String(required=True)
+        pass
